Fix array indexing and addressing, floats in arrays too
